+++ exercise_2_1/exercise_2_1.cpp
@@ -19,43 +19,6 @@ std::vector<double> readInVector(std::string s) {
   return result;
 }
 
-std::vector<double> applyConvolution(std::vector<double> x, std::vector<double> w, bool pack_with_zeros) {
-    std::vector<double> y;
-    int buf = (w.size()-1)/2;
-
-    // 2D loop to convolude vector x
-    for (int i = 0; i < x.size(); i++) {
-        double sum = 0;
-        for (int j = 0; j < w.size(); j++) {
-        // standard handling
-        if(i - buf + j >= 0 && i - buf + j < x.size()) {
-            sum += x[i - buf + j] * w[j];
-        } 
-        // underflow handling
-        else if (!pack_with_zeros && i - buf + j < 0) {
-            sum += x[0] * w[j];
-        }
-        // overflow handling
-        else if (!pack_with_zeros && i - buf + j >= x.size()) {
-            sum += x[x.size() -1] * w[j];
-        }
-        }
-        // store in output vector
-        y.push_back(sum);
-    }
-
-    // Print output vector
-    std::cout << "{";
-    for (int i = 0; i < y.size(); i++) {
-        std::cout << y[i] << (i != (y.size()-1) ? ", " : "");
-    }
-    std::cout << "}" << std::endl;
-
-    return y;
-  
-    // =========== END ===========
-}
-
 int main() {
     // get info from textfiles
     std::vector<double> x;
